back_end/requirements.txt
asgiref==3.8.1
cffi==1.17.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mysqlclient==2.2.7
pycparser==2.22
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.1
